\name{ACSWR-package}
\alias{ACSWR-package}
\alias{ACSWR}
\docType{package}
\title{
An R Companion Package for the Book "A Course in Statistics with R"
}
\description{
"A Course in Statistics with R" has been designed to meet the requirements of masters students.
}
\details{
\tabular{ll}{
Package: \tab ACSWR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-08-19\cr
License: \tab GPL-2 \cr
}
}
\author{
Prabhanjan Tattar

Maintainer: Prabhanjan Tattar <prabhanjannt@gmail.com>
}
\references{
Tattar, P. N., Suresh, R., and Manjunath, B. G. (2016). A Course in Statistics with R. J. Wiley. 
}
\keyword{Statistics}

\examples{
hist(rnorm(100))
}
