\name{kurtcoeff}
\alias{kurtcoeff}
\title{
Coefficient of Kurtosis
}
\description{
A simple function to obtain the coefficient of kurtosis on numeric variables.
}
\usage{
kurtcoeff(x)
}
\arguments{
  \item{x}{
the numeric vector for which the coefficient of kurtosis is required}
}
\details{
A straight-forward implementation of the formula is give here. A complete function "kurtosis" is available in the "e1071" package. 
}

\author{
Prabhanjan N. Tattar
}

\seealso{
e1071::kurtosis
}
\keyword{kurtosis}
\keyword{descriptive statistics}
\keyword{summary statistics}
