\name{LRNormalMean_UV}
\alias{LRNormalMean_UV}
\title{
Likelihood ratio test for mean when variance is unknown
}
\description{ Likelihood ratio test for mean when variance is unknown for a sample from normal distribution is setup here. 
}
\usage{
LRNormalMean_UV(x, mu0, alpha)
}
\arguments{
  \item{x}{
the variable of interest}
  \item{mu0}{
the mean value of interest
}
  \item{alpha}{
size of the LR test}
}

\author{
Prabhanjan N. Tattar}
\seealso{
LRNormalMean_KV
}

\keyword{Likelihood Ratio Test}
\keyword{Normal Distribution}
