\name{lval}
\alias{lval}
\title{
Letter Values}
\description{
This function is adapted from Prof. Jim Albert's "LearnEDA" package. It returns the letter values as discussed in Chapter 4. }
\usage{
lval(x, na.rm = TRUE)
}
\arguments{
  \item{x}{
the variable of interest
}
  \item{na.rm}{
the default setting removes the missing values
}
}
\author{
Prabhanjan Tattar 
}
\seealso{
LearnEDA
}
\keyword{Letter values}
\keyword{exploratory data analysis}
