\name{MPNormal}
\alias{MPNormal}
\title{
Most Powerful Test for Normal Distribution
}
\description{
The most powerful test for a sample from normal distribution is given here. The test is obtained by an application of the Neyman-Pearson lemma. 
}
\usage{
MPNormal(mu0, mu1, sigma, n, alpha)
}
\arguments{
  \item{mu0}{
mean under hypothesis H
}
  \item{mu1}{
mean under hypothesis K
}
  \item{sigma}{
standard deviation
}
  \item{n}{
sample size
}
  \item{alpha}{
size of the test
}
}
\author{
Prabhanjan N. Tattar
}
\seealso{
t.test
}
\keyword{Most Powerful Test}
\keyword{Normal Distribution}
