\name{MPPoisson}
\alias{MPPoisson}
\title{
Most Powerful Test for Poisson Distribution
}
\description{
The most powerful test for a sample from Poisson distribution is given here. The test is obtained by an application of the Neyman-Pearson lemma. 
 }
\usage{
MPPoisson(Hlambda, Klambda, alpha, n)
}
\arguments{
  \item{Hlambda}{
parameter under hypothesis H
}
  \item{Klambda}{
parameter under hypothesis K
}
  \item{alpha}{
size of the MP test
}
  \item{n}{
sample size
}
}

\author{
Prabhanjan N. Tattar
}
\keyword{Most Powerful Test}
\keyword{Poisson Distribution}
