\name{QH_CI}
\alias{QH_CI}
\title{
Quesenberry-Hurst Simultaneous Confidence Interval}
\description{
Quesenberry and Hurst (1964) have obtained the "simultaneous confidence intervals" for the vector of success in a multinomial distribution. 
}
\usage{
QH_CI(x, alpha)
}
\arguments{
  \item{x}{
a numeric vector }
  \item{alpha}{
as in 100 (1-alpha) }
}

\author{
Prabhanjan N. Tattar
}
\seealso{
prop.test
}
\keyword{Quesenberry-Hurst Simultaneous Confidence Interval}
