\name{resistant_line}
\alias{resistant_line}
\title{
Resistant Line EDA Regression Technique}
\description{
"Resistant Line" is an important EDA way of fitting a regression model. The function here develops the discussion in Section 4.5.1 Resistant Line. An alternative for this function is available in "rline" function of the "LearnEDA" package. 
}
\usage{
resistant_line(x, y, iterations)
}
\arguments{
  \item{x}{
the covariate or independent vector}
  \item{y}{
the dependent variate
}
  \item{iterations}{
the required number of iterations
}
}

\references{
Velleman, P.F., and Hoaglin, D.C. (2004). ABC of Exploratory Data Analysis.
Duxbury Press, Boston. Republished in 2004 by The Internet-First University
Press.
}
\author{
Prabhanjan N. Tattar
}
\seealso{
LearnEDA::rline
}
\keyword{EDA}
\keyword{Resistant Line}
