\name{skewcoeff}
\alias{skewcoeff}
\title{
A simple and straightforward function to compute the coefficient of skewness}
\description{
The function is fairly easy to follow. 
}
\usage{
skewcoeff(x)
}
\arguments{
  \item{x}{
variable of interest
}
}

\author{
Prabhanjan N. Tattar
}

\seealso{
e1071::skewness
}
\keyword{ skewness coefficient}
