\name{testtpm}
\alias{testtpm}
\docType{data}
\title{
A transition probability matrix}
\description{
A transition probaility matrix for understanding Markov chains. }
\usage{data(testtpm)}
\format{
A matrix of transition probability matrix
\describe{
    \item{\code{A}}{transitions probabilities from State A}
    \item{\code{B}}{transitions probabilities from State B}
    \item{\code{C}}{transitions probabilities from State C}
    \item{\code{D}}{transitions probabilities from State D}
    \item{\code{E}}{transitions probabilities from State E}
    \item{\code{F}}{transitions probabilities from State F}
  }
}
\examples{
data(testtpm)
}
\keyword{Markov chain}
