\name{testtpm3}
\alias{testtpm3}
\docType{data}
\title{
A matrix of transition probability matrix, third example
}
\description{
A transition probaility matrix for understanding Markov chains 
}
\usage{data(testtpm3)}
\format{
  A data frame with 7 observations on the following 7 variables.
  \describe{
    \item{\code{A}}{transitions probabilities from State A}
    \item{\code{B}}{transitions probabilities from State B}
    \item{\code{C}}{transitions probabilities from State C}
    \item{\code{D}}{transitions probabilities from State D}
    \item{\code{E}}{transitions probabilities from State E}
    \item{\code{F}}{transitions probabilities from State F}
    \item{\code{G}}{transitions probabilities from State G}
  }
}
\examples{
data(testtpm3)
}
\keyword{Markov chain}
