\name{TM}
\alias{TM}
\title{
Trimmed Mean}
\description{
The trimean can be viewed as the average of median and average of the lower and upper quartiles. A fairly simply function is defined here.}
\usage{
TM(x)
}
\arguments{
  \item{x}{
variable of interest
}
}
\author{
Prabhanjan N. Tattar
}
\seealso{
TMH, mean, median
}
\keyword{Trimean}
\keyword{EDA summary}