\name{TMH}
\alias{TMH}
\title{
Trimean based on hinges instead of quartiles}
\description{
The trimean is modified and defined based on hinges instead of the quartiles. }
\usage{
TMH(x)
}
\arguments{
  \item{x}{
variable of interest
}
}
\author{
Prabhanjan N. Tattar
}
\seealso{
TM
}
\keyword{Trimean based on hinges}
\keyword{EDA summary}