\name{UMPExponential}
\alias{UMPExponential}
\title{
Uniformly Most Powerful Test for Exponential Distribution}
\description{
A function is defined here which will return the uniformly most powerful test for exponential distribution. The function needs a simple use of the "qgamma" function. 
}
\usage{
UMPExponential(theta0, n, alpha)
}
\arguments{
  \item{theta0}{
the parameter of interest}
  \item{n}{
the sample size
}
  \item{alpha}{
level of the UMP test
}
}
\author{
Prabhanjan N. Tattar}
\keyword{Uniformly Most Powerful Test }
\keyword{Exponential Distribution}