\name{UMPNormal}
\alias{UMPNormal}
\title{
Uniformly Most Powerful Test for Normal Distribution
}
\description{
The "UMPNormal" function returns the critical points required for the UMP test for a sample from normal distribution. The standard deviation is assumed to be known. 
}
\usage{
UMPNormal(mu0, sigma, n, alpha)
}
\arguments{
  \item{mu0}{
the value of mean of interest}
  \item{sigma}{
standard deviation
}
  \item{n}{
sample size
}
  \item{alpha}{
size of the UMP test
}
}
\author{
Prabhanjan N. Tattar
}
\keyword{Uniformly Most Powerful Test}
\keyword{Normal distribution}