\name{UMPUniform}
\alias{UMPUniform}
\title{
Uniformly Most Powerful Test for Uniform Sample}
\description{
A simple and straightforward function for obtaining the UMP test for a random sample from uniform distribution. }
\usage{
UMPUniform(theta0, n, alpha)
}
\arguments{
  \item{theta0}{
the parameter value of interest
}
  \item{n}{
the sample size
}
  \item{alpha}{
the size of the required UMP test
}
}
\author{
Prabhanjan N. Tattar
}

\examples{
UMPUniform(0.6,10,0.05)
}
\keyword{Uniformly Most Powerful Test }
\keyword{Uniform Distribution}