\name{ww.test}
\alias{ww.test}
\title{
Wald-Wolfowitz Nonparametric Test
}
\description{
The "ww.test" function is an implementation of the famous Wald-Wolfowitz nonparametric test as discussed in Section 8.5. }
\usage{
ww.test(x, y)
}
\arguments{
  \item{x}{
values from sample 1
}
  \item{y}{
values from sample 2
}
}
\author{
Prabhanjan N. Tattar
}

\keyword{Wald-Wolfowitz }
\keyword{nonparametric test}